# Uses container/heap for DelayingQueue