--- a/staging/src/k8s.io/client-go/util/workqueue/rate_limitting_queue_test.go
+++ b/staging/src/k8s.io/client-go/util/workqueue/rate_limitting_queue_test.go
@@ -32,7 +32,7 @@ func TestRateLimitingQueue(t *testing.T) {
 		clock:           fakeClock,
 		heartbeat:       fakeClock.Tick(maxWait),
 		stopCh:          make(chan struct{}),
-		waitingForAddCh: make(chan waitFor, 1000),
+		waitingForAddCh: make(chan *waitFor, 1000),
 		metrics:         newRetryMetrics(""),
 	}
 	queue.DelayingInterface = delayingQueue