@deads2k 2017-05-10
I like the change, but I'm confused by what appears to be an inconsistent push, pop, and peek.