@alindeman 2017-05-11
Thanks for the review @deads2k.

The structure of the heap is a little confusing at first, I agree. But it is documented that the [minimum element is always at index 0](https://golang.org/pkg/container/heap):

> The minimum element in the tree is the root, at index 0.

`Push` and `Pop` do operate at the other 'end' of the heap, and `container/heap` will 'percolate' up or down the element into the correct location.

How do you best suggest I clarify this code? By example, do you mean an example test? I'd be happy to work one up if so.