@deads2k 2017-05-11
this lgtm

@liggitt want a look?