@alindeman 2017-05-11
Thanks for the review @liggitt. I've now added some documentation to the functions as well. What do you think?