@alindeman 2017-05-15
@deads2k Thanks, looks like we're fully green now. Let me know if you have any other feedback.